- fixed some styling
