+++ src/components/chart.tsx
@@ -10,7 +10,7 @@ import {
 } from 'chart.js';
 import { Line } from 'react-chartjs-2';
 import { useSelector } from 'react-redux';
-import { RootState } from './store';
+import { RootState } from '../store';
 
 ChartJS.register(
   CategoryScale,
