+++ src/visualize.tsx
@@ -1,21 +1,23 @@
 import { Button } from "antd";
 import { Link } from "react-router-dom";
-import { TransactionsChart } from "./chart";
+import { TransactionsChart } from "./components/chart";
 
 const Visualize: React.FC<any> = (props): JSX.Element => {
   return (
-    <>
-      <TransactionsChart />
+    <div className="chart">
+      <div className="chart-area">
+        <TransactionsChart />
+      </div>
 
       <br />
       <br />
       <br />
-      <Link to='/'>
+      <Link to='/' className="chart-button">
         <Button type="primary" htmlType="submit">
           Add more Transactions
         </Button>
       </Link>
-    </>
+    </div>
   );
 }
